+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.6
 - new buttons
+- new setting for visibility: always visible; old behavior; auto hide
 
 # 1.0.5
 - NUCLEAR and ghost options